optimize AST before printing
This introduces an `optimizeAst(…)` function that creates a fresh AST
from an AST but handles all the edge cases we used to have in
`toCss(…)`.

For example, `@property` is deduped (and fallbacks are generated for
older browsers), `Context` nodes are transparent and `AtRoots` move
nodes into the actual root.

This allows us to simplify the `toCss(…)` code to be a 1-to-1
translation and simply print `declarations`, `rules`, `at-rules` and
comments. We don't have to worry about the other special logic.

This also means that we don't have to re-introduce the same logic in our
PostCSS AST <-> Our AST transformations.

---

In addition: I ran some checks on the Catalyst codebase, and noticed
that we have the most declarations, then rules, then at-rules, then
at-roots, then context nodes and last but not least comments.

With this information in mind, the if-branches are using this
information to first check for declarations, then rules, ...